\name{bitterling}
\alias{bitterling}
\docType{data}
\title{Bitterling}
\description{
This dataset concerns reproductive behavior of male bitterlings with
data derived from 13 sequences using a moving time-window of size two. 
}
\usage{data(bitterling)}
\format{
  A frequency tables with bitterling reproductive behavior at time point 1 (rows)
  and at time point 2 (columns).
  \describe{
    \item{\code{jk}}{jerking}
    \item{\code{tu}}{turning beats}
    \item{\code{hb}}{head butting}
    \item{\code{chs}}{chasing}
    \item{\code{ft}}{fleeing}
    \item{\code{qu}}{quivering}
    \item{\code{le}}{leading}
    \item{\code{hdp}}{head down posture}
    \item{\code{sk}}{skimming}
    \item{\code{sn}}{snapping}
    \item{\code{chf}}{chafing}
    \item{\code{ffl}}{finflickering}
  }
}

\references{Wiepkema, P.R. (1961). An ethological analysis of the reproductive 
behavior of the bitterling (rhodeus amarus bloch). Archives Neerlandais Zoologique, 14, 103-199.
}
\examples{
data(bitterling)
}
\keyword{datasets}
