\name{galton}
\alias{galton}
\docType{data}
\title{Galton's RFF data}
\description{Records of family faculties cross-classification of midparent height 
and adult children heigth in inches.
}
\usage{data(galton)}
\format{
  A frequency table with 11 times 14 height classifications in inches.
}
\references{
Galton, F. (1889). Natural Inheritance. London: MacMillan. 
}
\examples{
data(galton)
}
\keyword{datasets}
