\name{srole}
\alias{srole}
\docType{data}
\title{Srole Data}
\description{This dataset provides a cross-classification
of subjects according to their mental health status and parents' socio-economic status.
}
\usage{data(srole)}
\format{
Mental health has four categories (rows): well, mild symptom formation, moderate symptom
formation, and impaired. There are six categories of socio-economic status in the columns. 
}

\references{Srole, L., Langner, T.S., Michael, S.T., Opler, M.K., & Rennie, T.A.C. (1962). 
Mental health in the metropolis: The midtown Manhattan study. New York: McGraw-Hill.
}
\examples{
data(srole)
## maybe str(srole) ; plot(srole) ...
}
\keyword{datasets}
