\name{tocher}
\alias{tocher}
\docType{data}
\title{Tocher's eye/hair color data.}
\description{Eye color and hair color cross-classification of 5387 Scottish school children.
}
\usage{data(tocher)}
\format{Frequency table with eye color in the rows (blue, light, medium, dark) and 
hair color in the columns (fair, red, medium, dark, black).
}
\references{Maung, K. (1941). Discriminant analysis of Tocher's eye colour data 
for Scottish school children. Annals of Eugenics, 11, 64-67.
}
\examples{
data(tocher)
## maybe str(tocher) ; plot(tocher) ...
}
\keyword{datasets}
